use policies in handlers

+++ src/routes/events/handlers.js
@@ -1,5 +1,7 @@
 import httpCodes from 'http-codes';
 import handleErrors from '../../lib/apiErrorHandler';
+import EventReport from '../../policies/event';
+import { currentUserId } from '../../services/currentUser';
 import {
   createEvent,
   findEventsByCollaboratorId,
@@ -10,10 +12,17 @@ import {
   updateEvent,
   destroyEvent,
 } from '../../services/event';
+import { userById } from '../../services/users';
 
 const namespace = 'SERVICE:EVENTS';
 
 const logContext = { namespace };
+
+export const getEventAuthorization = async (req, res, report) => {
+  const userId = await currentUserId(req, res);
+  const user = await userById(userId);
+  return new EventReport(user, report);
+};
 
 export const getHandler = async (req, res) => {
   try {
@@ -48,6 +57,12 @@ export const getHandler = async (req, res) => {
       return res.status(httpCodes.NOT_FOUND).send({ message: 'Event not found' });
     }
 
+    const auth = getEventAuthorization(req, res, event);
+
+    if (!auth.canRead()) {
+      return res.sendStatus(403);
+    }
+
     return res.status(httpCodes.OK).send(event);
   } catch (error) {
     return handleErrors(req, res, error, logContext);
@@ -59,6 +74,10 @@ export const createHandler = async (req, res) => {
     if (!req.body) {
       return res.status(httpCodes.BAD_REQUEST).send({ message: 'Request body is empty' });
     }
+
+    const { regionId } = req.body;
+    const auth = getEventAuthorization(req, res, { regionId });
+    if (!auth.canWriteInRegion()) { return res.sendStatus(403); }
 
     const event = await createEvent(req.body);
     return res.status(httpCodes.CREATED).send(event);
@@ -75,6 +94,10 @@ export const updateHandler = async (req, res) => {
       return res.status(httpCodes.BAD_REQUEST).send({ message: 'Request body is empty' });
     }
 
+    const { regionId } = req.body;
+    const auth = getEventAuthorization(req, res, { regionId });
+    if (!auth.canWriteInRegion()) { return res.sendStatus(403); }
+
     const event = await updateEvent(eventId, req.body);
     return res.status(httpCodes.CREATED).send(event);
   } catch (error) {
@@ -85,6 +108,15 @@ export const updateHandler = async (req, res) => {
 export const deleteHandler = async (req, res) => {
   try {
     const { eventId } = req.params;
+
+    const event = await findEventById(eventId);
+    if (!event) {
+      return res.status(httpCodes.NOT_FOUND).send({ message: 'Event not found' });
+    }
+
+    const auth = getEventAuthorization(req, res, event);
+    if (!auth.canDelete()) { return res.sendStatus(403); }
+
     await destroyEvent(eventId);
     return res.status(httpCodes.OK);
   } catch (error) {
