+++ src/routes/sessionReports/handlers.js
@@ -1,5 +1,8 @@
 import httpCodes from 'http-codes';
 import handleErrors from '../../lib/apiErrorHandler';
+import SessionReport from '../../policies/sessionReport';
+import { currentUserId } from '../../services/currentUser';
+import { findEventById } from '../../services/event';
 import {
   createSession,
   findSessionsByEventId,
@@ -7,10 +10,18 @@ import {
   updateSession,
   destroySession,
 } from '../../services/sessionReports';
+import { userById } from '../../services/users';
+import { getEventAuthorization } from '../events/handlers';
 
 const namespace = 'SERVICE:SESSIONREPORTS';
 
 const logContext = { namespace };
+
+export const getSessionAuthorization = async (req, res, report) => {
+  const userId = await currentUserId(req, res);
+  const user = await userById(userId);
+  return new SessionReport(user, report);
+};
 
 export const getHandler = async (req, res) => {
   try {
@@ -37,6 +48,12 @@ export const getHandler = async (req, res) => {
       return res.status(httpCodes.NOT_FOUND).send({ message: 'Session Report not found' });
     }
 
+    const auth = getSessionAuthorization(req, res, session);
+
+    if (!auth.canRead()) {
+      return res.sendStatus(403);
+    }
+
     return res.status(httpCodes.OK).send(session);
   } catch (error) {
     return handleErrors(req, res, error, logContext);
@@ -49,8 +66,20 @@ export const createHandler = async (req, res) => {
       return res.status(httpCodes.BAD_REQUEST).send({ message: 'Request body is empty' });
     }
 
-    const event = await createSession(req.body);
-    return res.status(httpCodes.CREATED).send(event);
+    const { eventId } = req.body;
+
+    if (eventId === undefined) {
+      return res.status(httpCodes.BAD_REQUEST).send({ message: 'Event ID is required' });
+    }
+
+    // Get associated event to use for authorization for region write
+    const event = await findEventById(eventId);
+    if (!event) { return res.status(httpCodes.NOT_FOUND).send({ message: 'Event not found' }); }
+    const auth = getEventAuthorization(req, res, event);
+    if (!auth.canWriteInRegion()) { return res.sendStatus(403); }
+
+    const session = await createSession(req.body);
+    return res.status(httpCodes.CREATED).send(session);
   } catch (error) {
     return handleErrors(req, res, error, logContext);
   }
@@ -64,8 +93,12 @@ export const updateHandler = async (req, res) => {
       return res.status(httpCodes.BAD_REQUEST).send({ message: 'Request body is empty' });
     }
 
-    const event = await updateSession(id, req.body);
-    return res.status(httpCodes.CREATED).send(event);
+    const session = findSessionById(id);
+    const auth = getSessionAuthorization(req, res, session);
+    if (!auth.canDelete()) { return res.sendStatus(403); }
+
+    const updatedSession = await updateSession(id, req.body);
+    return res.status(httpCodes.CREATED).send(updatedSession);
   } catch (error) {
     return handleErrors(req, res, error, logContext);
   }
@@ -74,6 +107,11 @@ export const updateHandler = async (req, res) => {
 export const deleteHandler = async (req, res) => {
   try {
     const { id } = req.params;
+
+    const session = findSessionById(id);
+    const auth = getSessionAuthorization(req, res, session);
+    if (!auth.canDelete()) { return res.sendStatus(403); }
+
     await destroySession(id);
     return res.status(httpCodes.OK);
   } catch (error) {
